refactor step5
