+++ src/main/java/carrace/domain/Car.java
@@ -1,4 +1,4 @@
-package carrace;
+package carrace.domain;
 
 import carrace.moving.RandomMoving;
