+++ src/main/java/carrace/domain/RaceWinner.java
@@ -1,4 +1,4 @@
-package carrace;
+package carrace.domain;
 
 import java.util.*;
 import java.util.stream.Collectors;
