+++ src/main/java/carrace/util/Splitter.java
@@ -1,6 +1,6 @@
 package carrace.util;
 
-import carrace.Car;
+import carrace.domain.Car;
 
 import java.util.List;
 import java.util.stream.Collectors;
